Implemented Program 6: Popular words
Chains 2 map-reduce jobs together.  takes 3 command line arguments
where the first is the input path, the second is the output path for
the first job and the input to the second job, and the third is the
final output path.  The final output’s keys are the count of the words
and the value is the word itself.  It’s sorted descending using a
Comparator class